Add files via upload
This is my first web page. I used HTML and CSS only. The only good feature it has is that it is responsive, no JS used so far. Next stop will be a better back-end to a new web page. Hope you like it!
If you are interested in visiting the web page, here is the link: https://laughing-montalcini-801eb4.netlify.app/